# docs/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "FlexPos.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5088/"
    }
  ],
  "paths": {
...
    "/api/ventas": {
      "get": {
        "tags": [
          "Ventas"
        ],
        "parameters": [
          {
            "name": "desdeUtc",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "hastaUtc",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "clienteId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "estado",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pagina",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "tamanoPagina",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PaginaVentasDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginaVentasDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginaVentasDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Ventas"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearVentaRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearVentaRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CrearVentaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VentaDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VentaDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VentaDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/ventas/{id}": {
      "get": {
        "tags": [
          "Ventas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VentaDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VentaDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VentaDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/ventas/{id}/pagos": {
      "post": {
        "tags": [
          "Ventas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgregarPagosVentaRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AgregarPagosVentaRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AgregarPagosVentaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VentaDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VentaDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VentaDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/ventas/{id}/devoluciones": {
      "post": {
        "tags": [
          "Ventas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevolucionVentaRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DevolucionVentaRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DevolucionVentaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VentaDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VentaDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VentaDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/ventas/{id}/anular": {
      "post": {
        "tags": [
          "Ventas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnularVentaRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AnularVentaRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AnularVentaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VentaDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VentaDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VentaDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/ventas/{id}/comprobante.pdf": {
      "get": {
        "tags": [
          "Ventas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/pdf": { }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/pdf": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/ventas/{id}/factura.pdf": {
      "get": {
        "tags": [
          "Ventas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/pdf": { }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/pdf": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AgregarPagosVentaRequest": {
        "required": [
          "pagos"
        ],
        "type": "object",
        "properties": {
          "pagos": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/PagoVentaRequest"
            }
          }
        }
      },
      "AnularVentaRequest": {
        "required": [
          "motivo",
          "pagos"
        ],
        "type": "object",
        "properties": {
          "motivo": {
            "type": [
              "null",
              "string"
            ]
          },
          "pagos": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/PagoVentaRequest"
            }
          }
        }
      },
...
      "CrearVentaRequest": {
        "required": [
          "clienteId",
          "solicitarFactura",
          "descuentoGeneral",
          "lineas",
          "pagos"
        ],
        "type": "object",
        "properties": {
          "clienteId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "solicitarFactura": {
            "type": "boolean"
          },
          "descuentoGeneral": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DescuentoVentaRequest"
              }
            ]
          },
          "lineas": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/LineaVentaRequest"
            }
          },
          "pagos": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/PagoVentaRequest"
            }
          }
        }
      },
      "DescuentoVentaRequest": {
        "required": [
          "tipo",
          "valor"
        ],
        "type": "object",
        "properties": {
          "tipo": {
            "type": [
              "null",
              "string"
            ]
          },
          "valor": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
...
      "DevolucionVentaRequest": {
        "required": [
          "motivo",
          "lineas",
          "pagos"
        ],
        "type": "object",
        "properties": {
          "motivo": {
            "type": [
              "null",
              "string"
            ]
          },
          "lineas": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/LineaDevolucionVentaRequest"
            }
          },
          "pagos": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/PagoVentaRequest"
            }
          }
        }
      },
      "DocumentoVentaDto": {
        "required": [
          "id",
          "tipoDocumento",
          "numeroCompleto",
          "fechaEmisionUtc",
          "total"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tipoDocumento": {
            "type": "string"
          },
          "numeroCompleto": {
            "type": "string"
          },
          "fechaEmisionUtc": {
            "type": "string",
            "format": "date-time"
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
...
      "LineaVentaRequest": {
        "required": [
          "tipo",
          "articuloInventarioId",
          "servicioId",
          "empleadoId",
          "cantidad",
          "descuento",
          "impuestoId"
        ],
        "type": "object",
        "properties": {
          "tipo": {
            "type": [
              "null",
              "string"
            ]
          },
          "articuloInventarioId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "servicioId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "empleadoId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "cantidad": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "descuento": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DescuentoVentaRequest"
              }
            ]
          },
          "impuestoId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
...
      "PaginaVentasDto": {
        "required": [
          "elementos",
          "pagina",
          "tamanoPagina",
          "totalElementos"
        ],
        "type": "object",
        "properties": {
          "elementos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VentaDto"
            }
          },
          "pagina": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "tamanoPagina": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalElementos": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "PagoVentaDto": {
        "required": [
          "id",
          "metodoPagoId",
          "metodoPagoNombre",
          "esEfectivo",
          "importe",
          "referencia",
          "fechaUtc"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "metodoPagoId": {
            "type": "string",
            "format": "uuid"
          },
          "metodoPagoNombre": {
            "type": "string"
          },
          "esEfectivo": {
            "type": "boolean"
          },
          "importe": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "referencia": {
            "type": [
              "null",
              "string"
            ]
          },
          "fechaUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PagoVentaRequest": {
        "required": [
          "metodoPagoId",
          "importe",
          "referencia"
        ],
        "type": "object",
        "properties": {
          "metodoPagoId": {
            "type": "string",
            "format": "uuid"
          },
          "importe": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "referencia": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "VentaDto": {
        "required": [
          "id",
          "cajaId",
          "clienteId",
          "clienteNombre",
          "clienteDocumento",
          "fechaVentaUtc",
          "codigoMoneda",
          "estado",
          "subtotal",
          "tipoDescuentoGeneral",
          "valorDescuentoGeneral",
          "descuentoLineas",
          "descuentoGeneral",
          "impuestos",
          "total",
          "totalPagado",
          "totalDevuelto",
          "totalReintegrado",
          "totalPendiente",
          "motivoAnulacion",
          "lineas",
          "pagos",
          "documentos",
          "devoluciones",
          "fechaCreacionUtc",
          "fechaModificacionUtc"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "cajaId": {
            "type": "string",
            "format": "uuid"
          },
          "clienteId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "clienteNombre": {
            "type": [
              "null",
              "string"
            ]
          },
          "clienteDocumento": {
            "type": [
              "null",
              "string"
            ]
          },
          "fechaVentaUtc": {
            "type": "string",
            "format": "date-time"
          },
          "codigoMoneda": {
            "type": "string"
          },
          "estado": {
            "type": "string"
          },
          "subtotal": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "tipoDescuentoGeneral": {
            "type": [
              "null",
              "string"
            ]
          },
          "valorDescuentoGeneral": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "descuentoLineas": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "descuentoGeneral": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "impuestos": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "totalPagado": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "totalDevuelto": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "totalReintegrado": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "totalPendiente": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "motivoAnulacion": {
            "type": [
              "null",
              "string"
            ]
          },
          "lineas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LineaVentaDto"
            }
          },
          "pagos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PagoVentaDto"
            }
          },
          "documentos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentoVentaDto"
            }
          },
          "devoluciones": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DevolucionVentaDto"
            }
          },
          "fechaCreacionUtc": {
            "type": "string",
            "format": "date-time"
          },
          "fechaModificacionUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      }
    }
  },
...
}